+++ src/ServiceBus/Helpers/PurgeOperationCompletedEventArgs.cs
@@ -1,0 +1,41 @@
+﻿#region Copyright
+//=======================================================================================
+// Microsoft Azure Customer Advisory Team 
+//
+// This sample is supplemental to the technical guidance published on my personal
+// blog at http://blogs.msdn.com/b/paolos/. 
+// 
+// Author: Paolo Salvatori
+//=======================================================================================
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// 
+// LICENSED UNDER THE APACHE LICENSE, VERSION 2.0 (THE "LICENSE"); YOU MAY NOT USE THESE 
+// FILES EXCEPT IN COMPLIANCE WITH THE LICENSE. YOU MAY OBTAIN A COPY OF THE LICENSE AT 
+// http://www.apache.org/licenses/LICENSE-2.0
+// UNLESS REQUIRED BY APPLICABLE LAW OR AGREED TO IN WRITING, SOFTWARE DISTRIBUTED UNDER THE 
+// LICENSE IS DISTRIBUTED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
+// KIND, EITHER EXPRESS OR IMPLIED. SEE THE LICENSE FOR THE SPECIFIC LANGUAGE GOVERNING 
+// PERMISSIONS AND LIMITATIONS UNDER THE LICENSE.
+//=======================================================================================
+#endregion
+
+using System;
+
+namespace ServiceBusExplorer.ServiceBus.Helpers
+{
+    public class PurgeOperationCompletedEventArgs : EventArgs
+    {
+        public bool IsDeadLetterQueue { get; set; }
+        public string EntityPath { get; set; }
+        public long ElapsedMilliseconds { get; set; }
+        public long TotalMessagesPurged { get; set; }
+
+        public PurgeOperationCompletedEventArgs(string entityPath, long elapsedMilliseconds, long totalMessagesPurged, bool isDeadLetterQueue)
+        {
+            this.EntityPath = entityPath;
+            this.ElapsedMilliseconds = elapsedMilliseconds;
+            this.TotalMessagesPurged = totalMessagesPurged;
+            this.IsDeadLetterQueue = isDeadLetterQueue;
+        }
+    }
+}
